- client: limit global prefs mod time to now - server: limit global prefs mod time to now     These changes address the situation where a server     sends out prefs with mod time far in the future,     and there's no way to undo them
svn path=/trunk/boinc/; revision=14664

+++ lib/prefs.h
@@ -130,7 +130,7 @@ public:
 
 
 struct GLOBAL_PREFS {
-    int mod_time;
+    double mod_time;
     bool run_on_batteries;
         // poorly named; what it really means is:
         // if false, suspend while on batteries
